Fix asset get endpoint
